refactor(project-create): factorize code that will be reused
