--- src/abstract-command.js
@@ -1,23 +1,0 @@
-const { Command } = require('@oclif/command');
-const Context = require('@forestadmin/context');
-
-const defaultPlan = require('./context/plan');
-
-class AbstractCommand extends Command {
-  init(plan) {
-    Context.init(plan || defaultPlan);
-    this.context = Context.inject();
-    // FIXME: Restore when no more Context.inject present in services.
-    // this.context = Context.execute(plan || defaultPlan);
-    const { assertPresent, logger, chalk } = this.context;
-    assertPresent({ logger, chalk });
-
-    /** @protected @readonly */
-    this.logger = logger;
-
-    /** @protected @readonly */
-    this.chalk = chalk;
-  }
-}
-
-module.exports = AbstractCommand;
