+++ src/commands/environments.js
@@ -5,20 +5,27 @@ const AbstractAuthenticatedCommand = require('../abstract-authenticated-command'
 const withCurrentProject = require('../services/with-current-project');
 
 class EnvironmentCommand extends AbstractAuthenticatedCommand {
-  init(plan) {
-    super.init(plan);
+  constructor(argv, config, plan) {
+    super(argv, config, plan);
     const { assertPresent, env, environmentsRenderer } = this.context;
     assertPresent({ env, environmentsRenderer });
     this.env = env;
     this.environmentsRenderer = environmentsRenderer;
   }
 
-  async runIfAuthenticated() {
+  async run() {
+    await this.checkAuthentication();
+
     const parsed = this.parse(EnvironmentCommand);
     const config = await withCurrentProject({ ...this.env, ...parsed.flags });
     const manager = new EnvironmentManager(config);
     const environments = await manager.listEnvironments();
     this.environmentsRenderer.render(environments, config);
+  }
+
+  async catch(error) {
+    await this.handleAuthenticationErrors(error);
+    return super.catch(error);
   }
 }
