+++ src/commands/environments/copy-layout.js
@@ -4,8 +4,8 @@ const EnvironmentManager = require('../../services/environment-manager');
 const withCurrentProject = require('../../services/with-current-project');
 
 class CopyLayoutCommand extends AbstractAuthenticatedCommand {
-  init(plan) {
-    super.init(plan);
+  constructor(argv, config, plan) {
+    super(argv, config, plan);
     const { assertPresent, chalk, env, inquirer } = this.context;
     assertPresent({
       chalk,
@@ -51,7 +51,9 @@ class CopyLayoutCommand extends AbstractAuthenticatedCommand {
     return { fromEnvironment, toEnvironment };
   }
 
-  async runIfAuthenticated() {
+  async run() {
+    await this.checkAuthentication();
+
     const oclifExit = this.exit.bind(this);
     const parsed = this.parse(CopyLayoutCommand);
     const config = await withCurrentProject({ ...this.env, ...parsed.flags, ...parsed.args });
@@ -112,6 +114,11 @@ class CopyLayoutCommand extends AbstractAuthenticatedCommand {
       }
       this.exit(1);
     }
+  }
+
+  async catch(error) {
+    await this.handleAuthenticationErrors(error);
+    return super.catch(error);
   }
 }
