+++ src/commands/environments/reset.js
@@ -7,15 +7,17 @@ const { handleError } = require('../../utils/error');
 const askForEnvironment = require('../../services/ask-for-environment');
 
 class ResetCommand extends AbstractAuthenticatedCommand {
-  init(plan) {
-    super.init(plan);
+  constructor(argv, config, plan) {
+    super(argv, config, plan);
     const { assertPresent, env, inquirer } = this.context;
     assertPresent({ env, inquirer });
     this.env = env;
     this.inquirer = inquirer;
   }
 
-  async runIfAuthenticated() {
+  async run() {
+    await this.checkAuthentication();
+
     const parsed = this.parse(ResetCommand);
     const envSecret = this.env.FOREST_ENV_SECRET;
     const commandOptions = { ...parsed.flags, ...parsed.args, envSecret };
@@ -64,6 +66,11 @@ class ResetCommand extends AbstractAuthenticatedCommand {
       }
       this.exit(2);
     }
+  }
+
+  async catch(error) {
+    await this.handleAuthenticationErrors(error);
+    return super.catch(error);
   }
 }
