+++ src/commands/init.js
@@ -24,8 +24,8 @@ const PROMPT_MESSAGE_AUTO_CREATING_ENV_FILE =
   'Do you want a new `.env` file (containing your environment variables) to be automatically created in your current folder?';
 
 class InitCommand extends AbstractAuthenticatedCommand {
-  init(plan) {
-    super.init(plan);
+  constructor(argv, config, plan) {
+    super(argv, config, plan);
     const { assertPresent, env, fs, inquirer, spinner } = this.context;
     assertPresent({
       env,
@@ -42,7 +42,9 @@ class InitCommand extends AbstractAuthenticatedCommand {
     this.environmentVariables = {};
   }
 
-  async runIfAuthenticated() {
+  async run() {
+    await this.checkAuthentication();
+
     try {
       this.spinner.start({ text: 'Selecting your project' });
       await this.spinner.attachToPromise(this.projectSelection());
@@ -155,6 +157,11 @@ class InitCommand extends AbstractAuthenticatedCommand {
       }
     }
     return displayEnvironmentVariablesAndCopyToClipboard(this.environmentVariables);
+  }
+
+  async catch(error) {
+    await this.handleAuthenticationErrors(error);
+    return super.catch(error);
   }
 }
