+++ src/commands/logout.js
@@ -1,14 +1,16 @@
 const AbstractCommand = require('../abstract-command');
 
 class LogoutCommand extends AbstractCommand {
-  init(plan) {
-    super.init(plan);
+  constructor(argv, config, plan) {
+    super(argv, config, plan);
     const { assertPresent, authenticator } = this.context;
     assertPresent({ authenticator });
     this.authenticator = authenticator;
   }
 
   async run() {
+    await this.checkAuthentication();
+
     await this.authenticator.logout({ log: true });
   }
 }
