+++ src/commands/switch.js
@@ -5,8 +5,8 @@ const ProjectManager = require('../services/project-manager');
 const withCurrentProject = require('../services/with-current-project');
 
 class SwitchCommand extends AbstractAuthenticatedCommand {
-  init(plan) {
-    super.init(plan);
+  constructor(argv, config, plan) {
+    super(argv, config, plan);
     const { assertPresent, env, inquirer } = this.context;
     assertPresent({ env, inquirer });
     this.env = env;
@@ -67,7 +67,9 @@ class SwitchCommand extends AbstractAuthenticatedCommand {
     return config;
   }
 
-  async runIfAuthenticated() {
+  async run() {
+    await this.checkAuthentication();
+
     try {
       const config = await this.getConfig();
       const branches = (await BranchManager.getBranches(config.envSecret)) || [];
@@ -101,6 +103,11 @@ class SwitchCommand extends AbstractAuthenticatedCommand {
       this.logger.error(customError);
       this.exit(2);
     }
+  }
+
+  async catch(error) {
+    await this.handleAuthenticationErrors(error);
+    return super.catch(error);
   }
 }
