+++ src/services/projects/create/project-creator.js
@@ -63,20 +63,11 @@ class ProjectCreator {
         authSecret: this.keyGenerator.generate(),
       };
     } catch (error) {
-      let message;
-      if (error.message === 'Unauthorized') {
-        message = `Your session has expired. Please log back in with the command \`${this.chalk.cyan(
-          'forest login',
-        )}\`.`;
-      } else if (error.message === 'Conflict') {
-        message = 'A project with this name already exists. Please choose another name.';
-      } else {
-        message = `${this.messages.ERROR_UNEXPECTED} ${this.chalk.red(error)}`;
+      if (error.message === 'Conflict') {
+        error.message = 'A project with this name already exists. Please choose another name.';
       }
 
-      return this.terminator.terminate(1, {
-        logs: [message],
-      });
+      throw error;
     }
   }
 }
