+++ test/commands/test-cli-helper/test-cli-command.js
@@ -1,6 +1,7 @@
+const { Config } = require('@oclif/config');
+
 const prepareCommand = ({ commandArgs = [], commandClass: CommandClass, commandPlan }) => {
-  const instance = new CommandClass(commandArgs);
-  instance.init(commandPlan);
+  const instance = new CommandClass(commandArgs, new Config({ root: process.cwd() }), commandPlan);
 
   return instance;
 };
